35 add email for contact user(buyer)
and change tabs on the page for product

+++ DALdv1/DALdv1/MegkomnatnyeDveri.cs
@@ -29,9 +29,9 @@ namespace DALdv1
         public string Cvet { get; set; }
         public string TypDveri { get; set; }
         public string VnytrenneeNapolnenie { get; set; }
-        public Nullable<int> Cena { get; set; }
+        public Nullable<decimal> Cena { get; set; }
         public Nullable<int> Skidka { get; set; }
-        public Nullable<int> CenaSoSkidkoy { get; set; }
+        public Nullable<decimal> CenaSoSkidkoy { get; set; }
         public string Opisanie { get; set; }
         public bool Publicaciya { get; set; }
         public string DopCharacteristics { get; set; }
