+++ Domain2/Domain2/Interfaces/IVhodnyeDvRepository.cs
@@ -11,7 +11,7 @@ namespace Domain2.Interfaces
     {
         IEnumerable<VhodnyeDveri> GetVhodnyeDv();
         VhodnyeDveri GetVhodnyeDvById(int id);
-        void CreateVhodnyeDv(int id, string naz, string pr, string strpr, string cvet, string napoln, string yplotnit, int? tmet, string furn, string petli, string osn, string ovn, int? tdpol, int? cena, int? skidka, int? csskid, string opis, bool publ, string dopchar);
+        void CreateVhodnyeDv(int id, string naz, string pr, string strpr, string cvet, string napoln, string yplotnit, double? tmet, string furn, string petli, string osn, string ovn, double? tdpol, decimal? cena, int? skidka, decimal? csskid, string opis, bool publ, string dopchar);
         void SaveVhodnyeDv(VhodnyeDveri temp);
         void DeleteVhodnyeDv(int id);
         IEnumerable<FotoVhodnyhDverey> GetFotoVhDv();
